third_party/blink/web_tests/virtual/streaming-declarative-shadow-dom-disabled/external/wpt/shadow-dom/declarative/declarative-shadow-dom-basic.tentative-expected.txt

This is a testharness.js-based test.
FAIL Declarative Shadow DOM: Basic test assert_equals: No leftover template node expected null but got Element node <template shadowrootmode="open">
    <slot id="s1" name="...
FAIL Declarative Shadow DOM: Feature detection assert_true: Unable to feature detect expected true got false
FAIL Shadowrootmode reflection assert_equals: The shadowRootMode IDL should reflect the content attribute expected (string) "open" but got (undefined) undefined
FAIL Declarative Shadow DOM: Fragment parser basic test assert_equals: No leftover template node expected null but got Element node <template shadowrootmode="open">
        <slot id="s1" na...
PASS Declarative Shadow DOM: Invalid shadow root attribute
FAIL Declarative Shadow DOM: Closed shadow root attribute assert_equals: No template - converted to shadow root expected null but got Element node <template shadowrootmode="closed">
      </template>
FAIL Declarative Shadow DOM: Missing closing tag assert_equals: No leftover template node expected null but got Element node <template shadowrootmode="open">
        <slot id="s1" na...
FAIL Declarative Shadow DOM: delegates focus attribute assert_true: No shadow root found expected true got false
FAIL Declarative Shadow DOM: Multiple roots assert_equals: No leftover template nodes from either root expected null but got Element node <template shadowrootmode="open">
    <span>root 1</span>
...
FAIL Declarative Shadow DOM: template containing declarative shadow root assert_true: Inner div should have a shadow root expected true got false
FAIL Declarative Shadow DOM: template containing (deeply nested) declarative shadow root assert_true: Inner div should have a shadow root expected true got false
FAIL Declarative Shadow DOM: template containing a template containing declarative shadow root assert_true: Inner div should have a shadow root expected true got false
FAIL Declarative Shadow DOM: template containing declarative shadow root and UA shadow root assert_true: Inner div should have a shadow root expected true got false
PASS Declarative Shadow DOM: template containing closed declarative shadow root and UA shadow root
PASS Declarative Shadow DOM: template root element
Harness: the test ran to completion.
